+++ sentry/src/main/java/io/sentry/Sentry.java
@@ -1362,71 +1362,67 @@ public final class Sentry {
   }
 
   /**
-   * @deprecated Use {@link #feedback()}.{@link IFeedbackApi#showForm() showForm()} instead.
+   * @deprecated Use {@link #feedback()}.{@link IFeedbackApi#show() show()} instead.
    */
   @Deprecated
   @SuppressWarnings("InlineMeSuggester")
   public static void showUserFeedbackForm() {
-    feedback().showForm();
+    feedback().show();
   }
 
   /**
    * @deprecated Use {@link #feedback()}.{@link
-   *     IFeedbackApi#showForm(SentryFeedbackOptions.OptionsConfigurator) showForm(configurator)}
-   *     instead.
+   *     IFeedbackApi#show(SentryFeedbackOptions.OptionsConfigurator) show(configurator)} instead.
    */
   @Deprecated
   @SuppressWarnings("InlineMeSuggester")
   public static void showUserFeedbackForm(
       final @Nullable SentryFeedbackOptions.OptionsConfigurator configurator) {
-    feedback().showForm(configurator);
+    feedback().show(configurator);
   }
 
   /**
-   * @deprecated Use {@link #feedback()}.{@link IFeedbackApi#showForm(SentryId,
-   *     SentryFeedbackOptions.OptionsConfigurator) showForm(associatedEventId, configurator)}
-   *     instead.
+   * @deprecated Use {@link #feedback()}.{@link IFeedbackApi#show(SentryId,
+   *     SentryFeedbackOptions.OptionsConfigurator) show(associatedEventId, configurator)} instead.
    */
   @Deprecated
   @SuppressWarnings("InlineMeSuggester")
   public static void showUserFeedbackForm(
       final @Nullable SentryId associatedEventId,
       final @Nullable SentryFeedbackOptions.OptionsConfigurator configurator) {
-    feedback().showForm(associatedEventId, configurator);
+    feedback().show(associatedEventId, configurator);
   }
 
   /**
-   * @deprecated Use {@link #feedback()}.{@link IFeedbackApi#showForm() showForm()} instead.
+   * @deprecated Use {@link #feedback()}.{@link IFeedbackApi#show() show()} instead.
    */
   @Deprecated
   @SuppressWarnings("InlineMeSuggester")
   public static void showUserFeedbackDialog() {
-    feedback().showForm();
+    feedback().show();
   }
 
   /**
    * @deprecated Use {@link #feedback()}.{@link
-   *     IFeedbackApi#showForm(SentryFeedbackOptions.OptionsConfigurator) showForm(configurator)}
-   *     instead.
+   *     IFeedbackApi#show(SentryFeedbackOptions.OptionsConfigurator) show(configurator)} instead.
    */
   @Deprecated
   @SuppressWarnings("InlineMeSuggester")
   public static void showUserFeedbackDialog(
       final @Nullable SentryFeedbackOptions.OptionsConfigurator configurator) {
-    feedback().showForm(configurator);
+    feedback().show(configurator);
   }
 
   /**
-   * @deprecated Use {@link #feedback()}.{@link IFeedbackApi#showForm(SentryId,
-   *     SentryFeedbackOptions.OptionsConfigurator) showForm(associatedEventId, configurator)}
-   *     instead.
+   * @deprecated Use {@link #feedback()}.{@link IFeedbackApi#show(SentryId,
+   *     SentryFeedbackOptions.OptionsConfigurator) show(associatedEventId, configurator)} instead.
    */
   @Deprecated
   @SuppressWarnings("InlineMeSuggester")
   public static void showUserFeedbackDialog(
       final @Nullable SentryId associatedEventId,
       final @Nullable SentryFeedbackOptions.OptionsConfigurator configurator) {
-    feedback().showForm(associatedEventId, configurator);
+    feedback().show(associatedEventId, configurator);
   }
 
   /**